Rename display -> _display
